Move AnalysisName and ExportData to separate components
