+++ src/components/Sidebar/AnalysisName/AnalysisName.css
@@ -1,0 +1,8 @@
+.analysis-name {
+  font-size: 18px;
+  font-style: italic;
+}
+
+.analysis-edit-form .input {
+  width: 100%;
+}
