+++ src/components/Sidebar/ExportData/index.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Segment, Header, Button } from 'semantic-ui-react'
+
+export default class ExportData extends React.PureComponent {
+  render () {
+    return (
+      <Segment>
+        <Header as="h3">Export</Header>
+        <Button icon="download" content="Download" color="blue" fluid />
+      </Segment>
+    )
+  }
+}
