+++ src/components/Sidebar/Sidebar.css
@@ -4,8 +4,3 @@
   overflow-x: hidden;
   overflow-y: auto;
 }
-
-.analysis-name {
-  font-size: 18px;
-  font-style: italic;
-}
